abstracted event validation and message extraction to own modules
