+++ src/extract-message-data.js
@@ -1,0 +1,9 @@
+const extractMessageData = (event) => {
+  try {
+    return JSON.parse(event.Records[0].Sns.Message)
+  } catch (e) {
+    throw new Error('Could not parse SNS message')
+  }
+}
+
+module.exports = extractMessageData
